requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.7.4
python-multipart==0.0.9
pydantic==2.9.2

httpx==0.27.2
